Added profiler config flag, and wrapped profiler changes within this. Moved profiler handleoverflow function to specific armv8 64 directory
