+++ config.cmake
@@ -405,6 +405,14 @@ config_option(
     DEFAULT_DISABLED OFF
 )
 
+config_option(
+    ProfilerEnable PROFILER_ENABLE
+    "Enable support for the seL4 Profiling framework. Forwards PMU events to userspace as PMUEvent faults, with \
+    relevant information contained within the message registers. "
+    DEPENDS "KernelBenchmarks"
+    DEFAULT OFF
+)
+
 config_choice(
     KernelBenchmarks
     KERNEL_BENCHMARK
